+++ luci-app-rakitanmanager/root/usr/share/rakitanmanager/plugins/adb-refresh-network.sh
@@ -7,50 +7,10 @@ else
 	ADBID="$1"
 fi
 
-#Delay reenabling network, default is 3
-if [ -z "$2" ]; then
-	echo "DELAY is unset, using default..."
-	DS="3"
-else
-	DS="$2"
-fi
-
-#Ping checker, default is 8.8.8.8
-if [ -z "$3" ]; then
-	echo "PINGCHECKER is unset, using default..."
-	PINGCK="8.8.8.8"
-else
-	PINGCK="$3"
-fi
-
-#Ping checker count, default is 10
-if [ -z "$4" ]; then
-	echo "PINGCOUNT is unset, using default..."
-	PINGCT="10"
-else
-	PINGCT="$4"
-fi
-
 for IPX in ${ADBID}
 do
-    #Ping checker system
-	echo "Checking connection from device [${IPX}] to [${PINGCK}] with ping count [${PINGCT}] times...."
-	httping -c ${PINGCT} ${PINGCK} > /tmp/anuping
-	if grep -q "100.00% failed" /tmp/anuping; then
-		echo "Network unavailable, restarting phone modem...."
-	else
-        #Force refresh network?
-        if [ "$5" == "force" ]; then
-        	echo "Network is available but Force refresh is set, restarting phone modem..."
-        else
-        	echo "Network is available but Force refresh is unset, leaving phone modem like default..."
-	        exit 0
-        fi
-	fi
-	
-    #Ping checker system
 	echo "Connecting to ${IPX} device..." 
-    echo "Airplane mode is at DISABLED state, will be enabled in ${DS} secs..."
+    echo "Airplane mode is at DISABLED state, will be enabled in 3 secs..."
     if [[ "$(adb -s ${IPX} shell settings get global airplane_mode_on)" == "0" ]]; then
     	adb -s "$IPX" settings put global airplane_mode_on 1 &>/dev/null
     	adb -s "$IPX" am broadcast -a android.intent.action.AIRPLANE_MODE --ez state true &>/dev/null
@@ -58,7 +18,7 @@ do
     if [[ "$(adb -s ${IPX} shell cmd connectivity airplane-mode)" == "disabled" ]]; then
     	adb -s "$IPX" shell cmd connectivity airplane-mode enable &>/dev/null
     fi
-    sleep "${DS}" &>/dev/null
+    sleep "3" &>/dev/null
 
     echo "Disabling airplane mode to get new IP and refreshed network..."
     if [[  "$(adb -s ${IPX} shell settings get global airplane_mode_on)" == "1" ]]; then